Replace duplicated keys
The primary motivation for this change is that I am trying to use
KMonad, which works on the key that is output, not the keycode. This
causes problems when the same key appears on the keyboard more than
once.

Also, there's few enough keys on an Ergodox. There's no point in
having the same keys duplicated on the base layer.